+++ src/MalScraper/Model/MainModel.php
@@ -146,6 +146,10 @@ class MainModel
                 $area = 'li a[href$=video]';
                 $additionalUrl = '?p='.$model->getPage();
                 break;
+            case 'EpisodeModel':
+                $area = 'li a[href$=episode]';
+                $additionalUrl = '?offset='. (100 * ($model->getPage() - 1));
+                break;
             case 'CharacterPeoplePictureModel':
                 $area = 'li a[href$=pictures]';
                 break;