+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### v1.4.2 -
 - `Add` Feature: Video list of anime
+- `Add` Feature: Episode list of anime
 - `Fix` Missing promotional video in `getInfo()`
 
 ### v1.4.1 - Jan 8, 2019